+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Changelog
 
-## 0.1.1 *(27.05.2016)*
+## 0.1.1 *(?)*
  * Adding support for overriding components and modules for Activities
+ * Parent is no longer required for views. App component will be used as default.
+ * Modules parameter is no longer required
+ * Depending on Dagger 2.4
 
 ## 0.1.0
  * Initial release